+++ packages/components/Checkbox/src/Checkbox.settings.ts
@@ -6,10 +6,11 @@ export const checkboxSelectActionLabel = 'Toggle the Checkbox';
 export const settings: IComposeSettings<ICheckboxType> = [
   {
     tokens: {
-      borderColor: 'menuItemText',
-      color: 'menuItemText',
-      backgroundColor: 'menuBackground',
+      backgroundColor: 'transparent',
       textBorderColor: 'transparent',
+      borderColor: 'checkboxBorderColor',
+      checkmarkColor: 'checkmarkColor',
+      checkmarkVisibility: 0,
     },
     root: {
       accessible: true,
@@ -19,7 +20,7 @@ export const settings: IComposeSettings<ICheckboxType> = [
         display: 'flex',
         alignItems: 'center',
         flexDirection: 'row',
-        minHeight: 14,
+        minHeight: 20,
         marginTop: 0,
         position: 'relative',
       },
@@ -28,27 +29,34 @@ export const settings: IComposeSettings<ICheckboxType> = [
       style: {
         borderStyle: 'solid',
         borderWidth: 1,
-        minHeight: 14,
-        minWidth: 14,
+        borderRadius: 2,
+        minHeight: 16,
+        minWidth: 16,
         marginEnd: 4,
       },
     },
     checkmark: {
       style: {
-        aspectRatio: 1,
         position: 'relative',
         opacity: 0,
         fontSize: 10,
-        textAlign: 'center',
-        textAlignVertical: 'center',
+        marginStart: 2,
         top: -1,
+      },
+    },
+    checkmarkIcon: {
+      width: 8,
+      height: 7,
+      style: {
+        marginVertical: 4.5,
+        marginHorizontal: 4,
       },
     },
     content: {
       variant: 'bodyStandard',
       style: {
         borderStyle: 'solid',
-        borderWidth: 2,
+        borderWidth: 1,
         marginTop: 3,
       },
     },
@@ -56,12 +64,16 @@ export const settings: IComposeSettings<ICheckboxType> = [
     _overrides: {
       focused: {
         tokens: {
-          backgroundColor: 'menuItemBackgroundHovered',
-          textBorderColor: 'focusBorder',
-          checkmarkColor: 'menuItemTextHovered',
+          backgroundColor: 'checkboxBackground',
+          checkmarkColor: 'checkboxBackground',
         },
       },
       checked: {
+        tokens: {
+          borderColor: 'checkboxBackground',
+          backgroundColor: 'checkboxBackground',
+          checkmarkVisibility: 1,
+        },
         checkmark: {
           style: {
             opacity: 1,
@@ -70,15 +82,18 @@ export const settings: IComposeSettings<ICheckboxType> = [
       },
       hovered: {
         tokens: {
-          backgroundColor: 'menuItemBackgroundHovered',
-          checkmarkColor: 'menuItemTextHovered',
+          backgroundColor: 'checkboxBackground',
+          checkmarkColor: 'checkboxBackground',
         },
       },
       disabled: {
         tokens: {
-          borderColor: 'buttonBorderDisabled',
-          color: 'disabledBodyText',
-          backgroundColor: 'background',
+          backgroundColor: 'checkboxBackgroundDisabled',
+        },
+        checkbox: {
+          style: {
+            opacity: 0.38,
+          },
         },
       },
       boxAtEnd: {
@@ -91,7 +106,7 @@ export const settings: IComposeSettings<ICheckboxType> = [
       },
       pressed: {
         tokens: {
-          backgroundColor: 'menuItemBackgroundPressed',
+          backgroundColor: 'checkboxBackground',
         },
       },
     },