+++ packages/components/ContextualMenu/src/Submenu.tsx
@@ -1,14 +1,7 @@
 /** @jsx withSlots */
 import * as React from 'react';
 import { View } from 'react-native';
-import {
-  submenuName,
-  SubmenuProps,
-  SubmenuSlotProps,
-  SubmenuType,
-  SubmenuRenderData,
-  SubmenuState
-} from './Submenu.types';
+import { submenuName, SubmenuProps, SubmenuSlotProps, SubmenuType, SubmenuRenderData, SubmenuState } from './Submenu.types';
 import { settings } from './Submenu.settings';
 import { IUseComposeStyling, compose } from '@uifabricshared/foundation-compose';
 import { useSelectedKey } from '@fluentui-react-native/interactive-hooks';
@@ -21,12 +14,7 @@ import { CMContext } from './ContextualMenu';
 export const Submenu = compose<SubmenuType>({
   displayName: submenuName,
   usePrepareProps: (userProps: SubmenuProps, useStyling: IUseComposeStyling<SubmenuType>) => {
-    const {
-      setShowMenu,
-      shouldFocusOnMount = true,
-      shouldFocusOnContainer = true,
-      ...rest
-    } = userProps;
+    const { setShowMenu, shouldFocusOnMount = true, shouldFocusOnContainer = true, ...rest } = userProps;
 
     // Grabs the context information from ContextualMenu (onDismissMenu callback)
     const context = React.useContext(CMContext);
@@ -34,24 +22,19 @@ export const Submenu = compose<SubmenuType>({
     // This hook updates the Selected Button and calls the customer's onClick function. This gets called after a button is pressed.
     const data = useSelectedKey(null, userProps.onItemClick);
 
-    const onShow = React.useCallback(
-      () => {
-        userProps ?.onShow();
-        context.isSubmenuOpen = true;
-      }, [context]
-    )
-    const onDismiss = React.useCallback(
-      () => {
-        userProps ?.onDismiss();
-        setShowMenu(false);
-        context.isSubmenuOpen = false;
-      }, [context, setShowMenu]
-    )
-    const dismissCallback = React.useCallback(
-      () => {
-        onDismiss();
-        context ?.onDismissMenu();
-      }, [onDismiss, context]);
+    const onShow = React.useCallback(() => {
+      userProps?.onShow();
+      context.isSubmenuOpen = true;
+    }, [context]);
+    const onDismiss = React.useCallback(() => {
+      userProps?.onDismiss();
+      setShowMenu(false);
+      context.isSubmenuOpen = false;
+    }, [context, setShowMenu]);
+    const dismissCallback = React.useCallback(() => {
+      onDismiss();
+      context?.onDismissMenu();
+    }, [onDismiss, context]);
 
     context.dismissSubmenu = onDismiss;
 
@@ -65,7 +48,7 @@ export const Submenu = compose<SubmenuType>({
         selectedKey: data.selectedKey,
         onItemClick: data.onKeySelect,
         onDismissMenu: dismissCallback,
-      }
+      },
     };
 
     const styleProps = useStyling(userProps, (override: string) => state[override] || userProps[override]);
@@ -75,14 +58,13 @@ export const Submenu = compose<SubmenuType>({
         ...rest,
         onShow: onShow,
         onDismiss: onDismiss,
-        setInitialFocus: shouldFocusOnMount
+        setInitialFocus: shouldFocusOnMount,
       },
       container: {
         accessible: shouldFocusOnContainer,
         focusable: shouldFocusOnContainer && containerFocus,
         onBlur: toggleContainerFocus,
-      }
-
+      },
     });
 
     return { slotProps, state };
@@ -90,22 +72,24 @@ export const Submenu = compose<SubmenuType>({
   settings: settings,
   slots: {
     root: Callout,
-    container: View
+    container: View,
   },
   styles: {
     root: [backgroundColorTokens, borderTokens],
-    container: []
+    container: [],
   },
   render: (Slots: ISlots<SubmenuSlotProps>, renderData: SubmenuRenderData, ...children: React.ReactNode[]) => {
     if (renderData.state == undefined) {
       return null;
     }
-    return<CMContext.Provider value={renderData.state.context}>
-      <Slots.root>
-        <Slots.container>{children}</Slots.container>
-      </Slots.root>
-    </CMContext.Provider>;
-  }
+    return (
+      <CMContext.Provider value={renderData.state.context}>
+        <Slots.root>
+          <Slots.container>{children}</Slots.container>
+        </Slots.root>
+      </CMContext.Provider>
+    );
+  },
 });
 
 export default Submenu;