+++ packages/experimental/use-styling/src/buildUseStyling.test.ts
@@ -127,7 +127,7 @@ describe('useStyling test suite', () => {
     const useStyling = buildUseStyling(baseOptions, themeHelper);
     const slotProps = useStyling({});
     const slotProps2 = useStyling({ p1: 2, p2: 'bar' });
-    Object.keys(slotProps).forEach(key => {
+    Object.keys(slotProps).forEach((key) => {
       expect(slotProps[key]).toBe(slotProps2[key]);
     });
   });