+++ packages/theming/theme-types/src/Color.types.ts
@@ -441,6 +441,10 @@ export interface ControlColorTokens {
   buttonHoveredSecondaryContent: ColorValue;
   buttonPressedSecondaryContent: ColorValue;
 
+  checkboxBackground: ColorValue;
+  checkboxBackgroundDisabled: ColorValue;
+  checkmarkColor: ColorValue;
+
   personaActivityRing: ColorValue;
   personaActivityGlow: ColorValue;
 }