+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Wed, 16 Jun 2021 21:31:58 GMT and should not be manually modified.
+This log was last generated on Wed, 23 Jun 2021 18:53:56 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.27.0
+
+Wed, 23 Jun 2021 18:53:56 GMT
+
+### Minor changes
+
+- Add new PersonaCoinFluentColor prop (ruaraki@microsoft.com)
+
+## 0.26.0
+
+Wed, 23 Jun 2021 17:46:02 GMT
+
+### Minor changes
+
+- Add tests for NativeDatePicker (chripear@microsoft.com)
+
+## 0.25.5
+
+Mon, 21 Jun 2021 11:19:28 GMT
+
+### Patches
+
+- Comment out tests that crash the app for iOS and macOS (#706) (tamasane@gmail.com)
+
+## 0.25.4
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Enable on Fluent Tester directory and run (ruaraki@microsoft.com)
 
 ## 0.25.3