+++ packages/theming/apple-theme/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/apple-theme
 
-This log was last generated on Wed, 16 Jun 2021 04:20:49 GMT and should not be manually modified.
+This log was last generated on Mon, 21 Jun 2021 11:19:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.2
+
+Mon, 21 Jun 2021 11:19:28 GMT
+
+### Patches
+
+- Add checkmarkColor to apple theme (tamasane@gmail.com)
+
+## 0.6.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to libraries core and theming, run (ruaraki@microsoft.com)
 
 ## 0.6.0